+++ app/models/his/his_hosxpv3.js
@@ -1242,14 +1242,14 @@ class HisHosxpv3Model {
             .leftJoin('ovst', 'referin.vn', 'ovst.vn')
             .leftJoin('refer_reply', 'referin.vn', 'refer_reply.vn')
             .leftJoin('doctor', 'ovst.doctor', 'doctor.code')
-            .select(db.raw(`'${hisHospcode}' as HOSPCODE`))
+            .select(db.raw(`? as HOSPCODE`, hisHospcode))
             .select('referin.refer_hospcode as HOSP_SOURCE', 'patient.cid as CID_IN', 'referin.hn as PID_IN', 'referin.vn as SEQ_IN', 'referin.docno as REFERID', 'referin.refer_date as DATETIME_REFER', 'referin.icd10 as detail', 'ovst.doctor as dr', 'doctor.licenseno as provider', 'refer_reply.diagnosis_text as reply_diagnostic', 'refer_reply.advice_text as reply_recommend')
-            .select(db.raw(`case when referin.referin_number then referin.referin_number else concat('${hisHospcode}-',referin.docno) end as REFERID_SOURCE`))
+            .select(db.raw(`case when referin.referin_number IS NOT NULL AND referin.referin_number !='' AND referin.referin_number !='-' then referin.referin_number else concat('${hisHospcode}-',referin.docno) end as REFERID_SOURCE`))
             .select(db.raw(`concat(refer_reply.reply_date, ' ',refer_reply.reply_time) as reply_date`))
             .select(db.raw(`'' as AN_IN, concat(referin.refer_hospcode,referin.referin_number) as REFERID_PROVINCE`))
             .select(db.raw(`concat(ovst.vstdate, ' ',ovst.vsttime) as DATETIME_IN, '1' as REFER_RESULT`))
             .select(db.raw(`concat(ovst.vstdate, ' ',ovst.vsttime) as D_UPDATE`))
-            .where(db.raw(`(referin.refer_date='${visitDate}' or referin.date_in='${visitDate}')`))
+            .where(db.raw(`(referin.refer_date=? or referin.date_in=?)`, [visitDate, visitDate]))
             .where(db.raw('length(referin.refer_hospcode)=5'))
             .whereNotNull('referin.vn')
             .whereNotNull('patient.hn')
@@ -1411,7 +1411,7 @@ class HisHosxpv3Model {
         }
         sql = sql.whereNotNull('ipt.ward')
             .whereNot('ipt.ward', '');
-        return sql.groupBy('ipt.ward').orderBy('ipt.ward');
+        return sql.groupBy(['ipt.ward', 'ward.name']).orderBy('ipt.ward');
     }
     concurrentIPDByClinic(db, date) {
         let sql = db('ipt')
@@ -1422,7 +1422,7 @@ class HisHosxpv3Model {
             .andWhere(function () {
             this.whereNull('ipt.dchdate').orWhere('ipt.dchdate', '>=', date);
         });
-        return sql.groupBy('ipt.spclty').orderBy('ipt.spclty');
+        return sql.groupBy(['ipt.spclty', 'clinic.name']).orderBy('ipt.spclty');
     }
     sumOpdVisitByClinic(db, date) {
         let sql = db('ovst')
@@ -1430,7 +1430,8 @@ class HisHosxpv3Model {
             .select('ovst.vstdate as date', 'spclty.nhso_code as cliniccode', 'spclty.name as clinicname', db.raw('SUM(CASE WHEN an IS NULL or an="" THEN 0 ELSE 1 END) AS admit'))
             .count('ovst.vstdate as cases')
             .where('ovst.vstdate', date);
-        return sql.groupBy('spclty.nhso_code').orderBy('spclty.nhso_code');
+        return sql.groupBy(['ovst.vstdate', 'spclty.nhso_code', 'spclty.name'])
+            .orderBy('spclty.nhso_code');
     }
 }
 exports.HisHosxpv3Model = HisHosxpv3Model;