+++ app/models/his/his_ssb.js
@@ -6,12 +6,13 @@ const hcode = process.env.HOSPCODE;
 const dbName = process.env.HIS_DB_NAME;
 const dbClient = process.env.HIS_DB_CLIENT;
 class HisSsbHModel {
-    async getAdmission(knex, columnName, searchNo, hospCode = hcode) {
-        columnName = columnName === 'visitNo' ? 'vn' : columnName;
-        return knex
-            .select('*')
-            .from('VW_PHER_ADMISSION')
-            .where(columnName, "=", searchNo);
+    async testConnect(db) {
+        let result;
+        result = await global.dbHIS('opdconfig').first();
+        const hospname = result?.hospitalname || result?.hospitalcode || null;
+        result = await db('patient').first();
+        const connection = result ? true : false;
+        return { hospname, connection };
     }
     getWard(db, wardCode = '', wardName = '') {
         let sql = db('lib_ward');
@@ -27,6 +28,21 @@ class HisSsbHModel {
         return sql
             .select('code as ward_code', 'ward as ward_name', 'standard as moph_code')
             .limit(maxLimit);
+    }
+    countBedNo(db) {
+        return { total_bed: 0 };
+    }
+    async getBedNo(db, bedno = null, start = -1, limit = 1000) {
+        return [];
+    }
+    concurrentIPDByWard(db, date) {
+        return [];
+    }
+    concurrentIPDByClinic(db, date) {
+        return [];
+    }
+    sumOpdVisitByClinic(db, date) {
+        return [];
     }
 }
 exports.HisSsbHModel = HisSsbHModel;