+++ app/plugins/db.js
@@ -49,12 +49,13 @@ const dbConnection = (type = 'HIS') => {
                 database: connection.database,
                 options: {
                     port: +connection.port,
-                    schema: connection.schema
+                    schema: connection.schema,
+                    trustServerCertificate: connection?.trustServerCertificate !== false
                 }
             }
         };
         if (connection?.encrypt) {
-            opt.connection.encrypt = connection.encrypt;
+            opt.connection.encrypt = connection?.encrypt === false ? false : 'strict';
         }
     }
     else if (config.client == 'oracledb') {