+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - ปรับ Source HIS: HI
   - เพิ่ม source HIS: MIT Net
   - ปรับ source HIS: HOSxP ให้ get clinic ที่ spclty แทน ipt_spclty
+  - เพิ่ม Group by แบบ Aggregate ใน Query ของ MOPH ERP
+  - ปรับการ connection mssql
   
 # 3.8.12
 ## Feature